wip: added initial conversion to snakeyaml
