+++ build.gradle.kts
@@ -27,7 +27,6 @@ plugins {
   id("com.gradle.plugin-publish") version "0.12.0"
 }
 
-apply(plugin = "io.spring.dependency-management")
 apply(from = "${project.rootDir}/gradle/versions.gradle")
 
 kotlinDslPluginOptions {
@@ -39,6 +38,7 @@ version = "0.1.2"
 
 repositories {
   jcenter()
+//  maven(url = "http://oss.sonatype.org/content/groups/public/")
 }
 
 gradlePlugin {
@@ -58,6 +58,12 @@ pluginBundle {
   tags = listOf("intellij", "generator", "kotlin-dsl")
 }
 
+publishing {
+  repositories {
+    maven(url = "build/repository")
+  }
+}
+
 tasks {
   withType<KotlinCompile>().configureEach {
     kotlinOptions.jvmTarget = "11"
@@ -69,20 +75,12 @@ tasks {
 }
 
 dependencies {
+  implementation("org.snakeyaml:snakeyaml-engine:2.1")
   implementation("org.redundent:kotlin-xml-builder")
-  implementation("com.fasterxml.jackson.core:jackson-databind")
-  implementation("com.fasterxml.jackson.module:jackson-module-kotlin")
-  implementation("com.fasterxml.jackson.dataformat:jackson-dataformat-yaml")
   implementation("javax.xml.bind:jaxb-api")
   implementation("com.sun.xml.bind:jaxb-core")
 
   testImplementation("io.kotest:kotest-runner-junit5-jvm")
   testImplementation("io.kotest:kotest-runner-console-jvm")
   testImplementation("io.kotest:kotest-assertions-core-jvm")
-}
-
-publishing {
-  repositories {
-    maven(url = "build/repository")
-  }
 }
