+++ src/main/kotlin/dev/north/fortyone/gradle/intellij/run/generator/models/IntellijRunConfig.kt
@@ -1,0 +1,5 @@
+package dev.north.fortyone.gradle.intellij.run.generator.models
+
+interface IntellijRunConfig {
+  fun toXml(): String
+}
