+++ src/test/resources/gradle-config.yaml
@@ -1,0 +1,6 @@
+--- !!dev.north.fortyone.gradle.intellij.run.generator.models.GradleRunConfig
+name: GRADLE | Ktlint > Format
+filename: GRADLE__Ktlint__Format.xml
+default: false
+tasks:
+  - ktlintFormat
