+++ src/test/resources/multiple-definitions.yaml
